// src/types/rick-and-morty.ts
export interface RickAndMortyListItem {
  id: number;
  name: string;
  status: string;
  species: string;
  type: string;
  gender: string;
  origin: {
    name: string;
    url: string;
  };
  location: {
    name: string;
    url: string;
  };
  image: string;
  episode: string[];
  url: string;
  created: string;
}

export interface RickAndMortyInfo {
  count: number;
  pages: number;
  next: string | null;
  prev: string | null;
}

export interface RickAndMortyListResponse {
  info: RickAndMortyInfo;
  results: RickAndMortyListItem[];
}

export type RickAndMortyCharacter = RickAndMortyListItem;